Remove unnecessary extensions.
