+++ PlanetLite/AppWindowController.swift
@@ -200,3 +200,16 @@ extension AppWindowController: NSToolbarDelegate {
     }
 
 }
+
+
+extension NSToolbar.Identifier {
+    static let toolbarIdentifier = NSToolbar.Identifier("PlanetLiteWindowToolbar")
+}
+
+
+extension NSToolbarItem.Identifier {
+    static let sidebarSeparatorItem = NSToolbarItem.Identifier("PlanetLiteToolbarSidebarSeparatorItem")
+    static let sidebarItem = NSToolbarItem.Identifier("PlanetLiteToolbarSidebarItem")
+    static let addItem = NSToolbarItem.Identifier("PlanetLiteToolbarAddItem")
+    static let showInfoItem = NSToolbarItem.Identifier("PlanetLiteToolbarShowInfoItem")
+}
